Inserting Data

+++ book_store/book_outlet/models.py
@@ -5,4 +5,4 @@ from django.db import models
 
 class Book(models.Model):
     title = models.CharField(max_length=50)
-    rating = models.PositiveIntegerField()
+    rating = models.DecimalField(max_digits=3, decimal_places=2)
